adj. pre

+++ preprocess.py
@@ -75,7 +75,7 @@ def process_dataset(image_seg_pairs, root_path, quant_path, transposing, crop_in
         df2 = df[df['sample_id'] == sample_id].copy()
         df2 = df2[['cell_id', 'cell_type']]
         max_ids = segmask.max()
-        labels_array = np.full(max_ids, -1, dtype=float)
+        labels_array = np.full(max_ids, -1, dtype=int)
         for _, row in df2.iterrows():
             cell_id = int(row['cell_id'])
             if 1 <= cell_id <= max_ids:
